FIX: max-switch-actions constraint
The max-switch-actions constraint was broken due to the max_switch_actions data not getting properly passed to the math model upon transformation. This required a fix in PMD, and update to new release v0.11.5 for a min version

This also adds a :max_switch_actions ref which is Inf if the data field is not present.

This also updates the constraint formulation since in reality the constraint should be sum(|z - z_start|) <= max_switch_actions, and it was not formulated correctly before.