FIX: Powerflow output missing components
Because components whose status is DISABLED are excluded from variable creation, they do not appear in the solution, which means that the items in the solution from step to step may not be the same.

This fixes the issue by defaulting to zeros if the item is DISABLED, so that every object appears in every timestep.

Requires an update to PowerModelsDistribution v0.11.10