FIX: missing protection devices from fault current output
This fixes a problem with the fault current outputs where not all
protection devices were included in the output. Specifically,
when the protection devices were monitoring non-switch branches
they were not being included. This updates the output to include
line objects as well, with the same schema as switch object have.